Move python visualization files
