+++ Python/tigre/__init__.py
@@ -3,8 +3,9 @@ from .utilities.geometry import geometry
 from .utilities.geometry_default import ConeGeometryDefault as geometry_default
 from .utilities.Ax import Ax
 from .utilities.Atb import Atb
-from .utilities.plotproj import plotproj, plotProj, plotSinogram
-from .utilities.plotimg import plotimg, plotImg
-from .utilities.plot_geometry import plot_geometry
+from .utilities.visualization.plotproj import plotproj, plotProj, plotSinogram
+from .utilities.visualization.plotimg import plotimg, plotImg
+from .utilities.visualization.plot_geometry import plot_geometry
+from .utilities.visualization.plot_angles import plot_angles
 from .utilities.CTnoise import add
 from . import algorithms
